refactor(authz-ui): move entity data types into lib/, drop unused category metadata
EntityInstance/AttrValue/EntitySource move to lib/entity.types.ts so the
data layer no longer imports from app/features/. Category metadata
(CATEGORIES, getCategory, getEntityCategoryId) had no consumers in this
slice and is deferred to the pages PR.
