+++ gravitee-gamma/gravitee-gamma-module-authz/src/main/ui/lib/__tests__/policy-entity-refs.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { describe, expect, it } from 'vitest';
-import type { EntityInstance } from '../../app/features/policy-structure/entity-types';
+import type { EntityInstance } from '../entity.types';
 import type { PolicyResponse } from '../api/authz-api.types';
 import { buildPolicyEntityRefs, extractEntityRefsFromPolicyText, type PolicyRef } from '../policy-entity-refs';
